# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于Gin的后台管理系统接口文档",
        "title": "Go-admin 后台管理系统",
        "contact": {},
        "version": "2.0.0"
    },
    "host": "localhost:8888",
    "basePath": "/admin-api/v1",
    "paths": {
...
        "/admin/sys/sys-dict/type": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取字典类型分页列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统字典管理"
                ],
                "summary": "获取字典类型分页列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典名称",
                        "name": "dictName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典类型",
                        "name": "dictType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建开始时间",
                        "name": "beginCreatedAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "创建结束时间",
                        "name": "endCreatedAt",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "新增字典类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统字典管理"
                ],
                "summary": "新增字典类型",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SysDictTypeInsertReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除字典类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统字典管理"
                ],
                "summary": "删除字典类型",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SysDictrDeleteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/sys/sys-dict/type/all-with-data": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取字典类型全部列表(关联字典数据)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统字典管理"
                ],
                "summary": "获取字典类型全部列表(关联字典数据)",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/sys/sys-dict/type/export": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "导出字典类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "系统字典管理"
                ],
                "summary": "导出字典类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典名称",
                        "name": "dictName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典类型",
                        "name": "dictType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/sys/sys-dict/type/option-select": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取字典类型全部列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统字典管理"
                ],
                "summary": "获取字典类型全部列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "字典名称",
                        "name": "dictName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "字典类型",
                        "name": "dictType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/sys/sys-dict/type/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取字典类型详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统字典管理"
                ],
                "summary": "获取字典类型详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新字典类型",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统字典管理"
                ],
                "summary": "更新字典类型",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "字典类型编号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SysDictTypeUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.SysDictTypeInsertReq": {
            "type": "object",
            "properties": {
                "dictName": {
                    "type": "string"
                },
                "dictType": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "dto.SysDictTypeUpdateReq": {
            "type": "object",
            "properties": {
                "dictName": {
                    "type": "string"
                },
                "dictType": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                }
            }
        },
        "dto.SysDictrDeleteReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "requestId": {
                    "description": "数据集",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}